Add support for 128-bit TraceId (6to7)  (#4918)

+++ dd-trace-core/src/test/groovy/datadog/trace/core/propagation/HttpExtractorTest.groovy
@@ -52,7 +52,7 @@ class HttpExtractorTest extends DDSpecification {
       if (expectedTraceId == null) {
         assert context == null
       } else {
-        assert context.traceId == DDTraceId.from(expectedTraceId)
+        assert context.traceId.toLong() == DDTraceId.from(expectedTraceId).toLong()
         assert context.spanId == DDSpanId.from(expectedSpanId)
       }
     }
